Mobile_Robot_StableValue_Wall_Following
We implement by using vscode, it bases on Python to control simulator in CoppeliaSim (V-REP). The mission of assignment is randomly the robot and detect the wall after then the robot will run nearly the wall. The achievement is how to control the speed of robot, turn left or right, how to use ultrasonic sensors.